Refactor proxy config/options in anticipation of mqtt refactor
