+++ include/aws/http/connection.h
@@ -122,98 +122,6 @@ struct aws_http_connection_monitoring_options {
 };
 
 /**
- * @Deprecated - Supported proxy authentication modes.  Superceded by proxy strategy.
- */
-enum aws_http_proxy_authentication_type {
-    AWS_HPAT_NONE = 0,
-    AWS_HPAT_BASIC,
-};
-
-/**
- * Supported proxy connection types
- */
-enum aws_http_proxy_connection_type {
-    /**
-     * Deprecated, but 0-valued for backwards compatibility
-     *
-     * If tls options are provided (for the main connection) then treat the proxy as a tunneling proxy
-     * If tls options are not provided (for the main connection), then treat the proxy as a forwarding proxy
-     */
-    AWS_HPCT_HTTP_LEGACY = 0,
-
-    /**
-     * Use the proxy to forward http requests.  Attempting to use both this mode and TLS on the tunnel destination
-     * is a configuration error.
-     */
-    AWS_HPCT_HTTP_FORWARD,
-
-    /**
-     * Use the proxy to establish an http connection via a CONNECT request to the proxy.  Works for both plaintext and
-     * tls connections.
-     */
-    AWS_HPCT_HTTP_TUNNEL,
-};
-
-struct aws_http_proxy_strategy;
-
-/**
- * Options for http proxy server usage
- */
-struct aws_http_proxy_options {
-
-    /**
-     * Type of proxy connection to make
-     */
-    enum aws_http_proxy_connection_type connection_type;
-
-    /**
-     * Proxy host to connect to
-     */
-    struct aws_byte_cursor host;
-
-    /**
-     * Port to make the proxy connection to
-     */
-    uint16_t port;
-
-    /**
-     * Optional.
-     * TLS configuration for the Local <-> Proxy connection
-     * Must be distinct from the the TLS options in the parent aws_http_connection_options struct
-     */
-    const struct aws_tls_connection_options *tls_options;
-
-    /**
-     * Optional
-     * Advanced option that allows the user to create a custom strategy that gives low-level control of
-     * certain logical flows within the proxy logic.
-     *
-     * For tunneling proxies it allows custom retry and adaptive negotiation of CONNECT requests.
-     * For forwarding proxies it allows custom request transformations.
-     * Other proxy connection types TBD.
-     */
-    struct aws_http_proxy_strategy *proxy_strategy;
-
-    /**
-     * @Deprecated - What type of proxy authentication to use, if any.
-     * Replaced by proxy_strategy
-     */
-    enum aws_http_proxy_authentication_type auth_type;
-
-    /**
-     * @Deprecated - Optional user name to use for basic authentication
-     * Replaced by proxy_strategy, using an instantiation from aws_http_proxy_strategy_new_basic_auth()
-     */
-    struct aws_byte_cursor auth_username;
-
-    /**
-     * @Deprecated - Optional password to use for basic authentication
-     * Replaced by proxy_strategy, using an instantiation from aws_http_proxy_strategy_new_basic_auth()
-     */
-    struct aws_byte_cursor auth_password;
-};
-
-/**
  * Options specific to HTTP/1.x connections.
  * Initialize with AWS_HTTP1_CONNECTION_OPTIONS_INIT to set default values.
  */
